+++ dev/src/qmaserv/cserv/linux.h
@@ -1,5 +1,6 @@
 #include <stdio.h>
+#include <stdint.h>
 
 int flip4(int iToFlip);
 short flip2(short shToFlip);
-int flip4array(long* in,short bytes);
+int flip4array(int32_t* in,short bytes);
